Added boolean to avoid multiple errors
https://github.com/balderdashy/skipper/issues/49
Adapter is emitting multiple errors on aborted upload. After the first one is emitted, the underlying skipper finishes working and the subsequent errors cause the server to crash. We still need to check if this is an issue with this adapter or with the overall architecture of skipper (that should have mechanisms to deal with a misbehaving adapter)